pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deeplens"
version = "0.1.0"
description = "Multi-agent entity research system built with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "openai>=1.0",
    "google-api-python-client>=2.0",
    "tavily-python>=0.5",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "streamlit>=1.30",
    "typer>=0.12",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
deeplens = "deeplens.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/deeplens"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
